Generalizando Todos los Cambios desde Formulario de Registrar hasta accesos de login en navegador y controlador de docente con sus vistas
