+++ Saaloon/Saaloon/Context/DBPortalEdu.dbml
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?><Database Name="PortalEdu" Class="DBPortalEduDataContext" xmlns="http://schemas.microsoft.com/linqtosql/dbml/2007">
-  <Connection Mode="WebSettings" ConnectionString="Data Source=DESKTOP-JH40SFF\SQLEXPRESS;Initial Catalog=PortalEdu;Integrated Security=True" SettingsObjectName="System.Configuration.ConfigurationManager.ConnectionStrings" SettingsPropertyName="PortalEduConnectionString1" Provider="System.Data.SqlClient" />
+  <Connection Mode="WebSettings" ConnectionString="Data Source=.;Initial Catalog=PortalEdu;Integrated Security=True" SettingsObjectName="System.Configuration.ConfigurationManager.ConnectionStrings" SettingsPropertyName="PortalEduConnectionString" Provider="System.Data.SqlClient" />
   <Table Name="dbo.Alumno" Member="Alumno">
     <Type Name="Alumno">
       <Column Name="IdAlumno" Type="System.Int32" DbType="Int NOT NULL IDENTITY" IsPrimaryKey="true" IsDbGenerated="true" CanBeNull="false" />
@@ -121,57 +121,6 @@
       <Association Name="Usuario_Docentes" Member="Docentes" ThisKey="IdUsuario" OtherKey="idUsuario" Type="Docentes" />
     </Type>
   </Table>
-  <Table Name="dbo.Inscripciones_del_Curso" Member="Inscripciones_del_Curso">
-    <Type Name="Inscripciones_del_Curso">
-      <Column Name="[Nombre del Alumno]" Member="Nombre_del_Alumno" Type="System.String" DbType="VarChar(30)" CanBeNull="true" />
-      <Column Name="[Apellido del Alumno]" Member="Apellido_del_Alumno" Type="System.String" DbType="VarChar(30)" CanBeNull="true" />
-      <Column Name="Curso" Type="System.String" DbType="VarChar(30)" CanBeNull="true" />
-    </Type>
-  </Table>
-  <Table Name="dbo.ViewAlumnos" Member="ViewAlumnos">
-    <Type Name="ViewAlumnos">
-      <Column Name="IdAlumno" Type="System.Int32" DbType="Int NOT NULL" CanBeNull="false" />
-      <Column Name="Nombre" Type="System.String" DbType="VarChar(30)" CanBeNull="true" />
-      <Column Name="Apellido" Type="System.String" DbType="VarChar(30)" CanBeNull="true" />
-      <Column Name="[Fecha de Nacimiento]" Member="Fecha_de_Nacimiento" Type="System.DateTime" DbType="DateTime" CanBeNull="true" />
-      <Column Name="Género" Type="System.Char" DbType="Char(1)" CanBeNull="true" />
-      <Column Name="Usuario" Type="System.String" DbType="VarChar(60)" CanBeNull="true" />
-      <Column Name="Correo" Type="System.String" DbType="VarChar(60)" CanBeNull="true" />
-      <Column Name="[Tipo Usuario]" Member="Tipo_Usuario" Type="System.Int32" DbType="Int" CanBeNull="true" />
-    </Type>
-  </Table>
-  <Table Name="dbo.ViewCurso" Member="ViewCurso">
-    <Type Name="ViewCurso">
-      <Column Name="IdCurso" Type="System.Int32" DbType="Int NOT NULL" CanBeNull="false" />
-      <Column Name="Nombre" Type="System.String" DbType="VarChar(30)" CanBeNull="true" />
-      <Column Name="Descripción" Type="System.String" DbType="VarChar(200)" CanBeNull="true" />
-      <Column Name="Recursos" Type="System.String" DbType="VarChar(30)" CanBeNull="true" />
-      <Column Name="Costo" Type="System.Decimal" DbType="Money" CanBeNull="true" />
-      <Column Name="[Nombre Docente]" Member="Nombre_Docente" Type="System.String" DbType="VarChar(30)" CanBeNull="true" />
-    </Type>
-  </Table>
-  <Table Name="dbo.ViewDocentes" Member="ViewDocentes">
-    <Type Name="ViewDocentes">
-      <Column Name="IdDocente" Type="System.Int32" DbType="Int NOT NULL" CanBeNull="false" />
-      <Column Name="Nombre" Type="System.String" DbType="VarChar(30)" CanBeNull="true" />
-      <Column Name="Apellido" Type="System.String" DbType="VarChar(30)" CanBeNull="true" />
-      <Column Name="[Fecha de Nacimiento]" Member="Fecha_de_Nacimiento" Type="System.DateTime" DbType="DateTime" CanBeNull="true" />
-      <Column Name="Género" Type="System.Char" DbType="Char(1)" CanBeNull="true" />
-      <Column Name="Usuario" Type="System.String" DbType="VarChar(60)" CanBeNull="true" />
-      <Column Name="Correo" Type="System.String" DbType="VarChar(60)" CanBeNull="true" />
-      <Column Name="[Tipo Usuario]" Member="Tipo_Usuario" Type="System.Int32" DbType="Int" CanBeNull="true" />
-    </Type>
-  </Table>
-  <Table Name="dbo.Temario" Member="Temario">
-    <Type Name="Temario">
-      <Column Name="IdTema" Type="System.Int32" DbType="Int NOT NULL IDENTITY" IsPrimaryKey="true" IsDbGenerated="true" CanBeNull="false" />
-      <Column Name="Tema" Type="System.String" DbType="VarChar(30)" CanBeNull="true" />
-      <Column Name="Descripcion" Type="System.String" DbType="VarChar(200)" CanBeNull="true" />
-      <Column Name="FotoTema" Type="System.String" DbType="VarChar(MAX)" CanBeNull="true" />
-      <Column Name="IdCurso" Type="System.Int32" DbType="Int" CanBeNull="true" />
-      <Association Name="Cursos_Temario" Member="Cursos" ThisKey="IdCurso" OtherKey="IdCurso" Type="Cursos" IsForeignKey="true" />
-    </Type>
-  </Table>
   <Function Name="dbo.SP_ModificaAlumno" Method="SP_ModificaAlumno">
     <Parameter Name="idAlumno" Type="System.Int32" DbType="Int" />
     <Parameter Name="nombre" Type="System.String" DbType="VarChar(30)" />
@@ -190,6 +139,16 @@
     <Parameter Name="Costo" Parameter="costo" Type="System.Decimal" DbType="Money" />
     <Parameter Name="Foto" Parameter="foto" Type="System.String" DbType="VarChar(MAX)" />
     <ElementType Name="SP_ModificaCursosResult">
+      <Column Name="Mensaje" Type="System.String" DbType="VarChar(31) NOT NULL" CanBeNull="false" />
+    </ElementType>
+  </Function>
+  <Function Name="dbo.SP_ModificaDocente" Method="SP_ModificaDocente">
+    <Parameter Name="idDocentes" Type="System.Int32" DbType="Int" />
+    <Parameter Name="nombre" Type="System.String" DbType="VarChar(30)" />
+    <Parameter Name="apellido" Type="System.String" DbType="VarChar(30)" />
+    <Parameter Name="fecha_n" Type="System.DateTime" DbType="DateTime" />
+    <Parameter Name="genero" Type="System.Char" DbType="Char(1)" />
+    <ElementType Name="SP_ModificaDocenteResult">
       <Column Name="Mensaje" Type="System.String" DbType="VarChar(31) NOT NULL" CanBeNull="false" />
     </ElementType>
   </Function>
@@ -217,16 +176,6 @@
     <Parameter Name="Activo" Parameter="activo" Type="System.Int32" DbType="Int" />
     <Parameter Name="tipo" Type="System.Int32" DbType="Int" />
     <ElementType Name="SP_ModificaUsuarioResult">
-      <Column Name="Mensaje" Type="System.String" DbType="VarChar(31) NOT NULL" CanBeNull="false" />
-    </ElementType>
-  </Function>
-  <Function Name="dbo.SP_ModificaDocente" Method="SP_ModificaDocente">
-    <Parameter Name="idDocentes" Type="System.Int32" DbType="Int" />
-    <Parameter Name="nombre" Type="System.String" DbType="VarChar(30)" />
-    <Parameter Name="apellido" Type="System.String" DbType="VarChar(30)" />
-    <Parameter Name="fecha_n" Type="System.DateTime" DbType="DateTime" />
-    <Parameter Name="genero" Type="System.Char" DbType="Char(1)" />
-    <ElementType Name="SP_ModificaDocenteResult">
       <Column Name="Mensaje" Type="System.String" DbType="VarChar(31) NOT NULL" CanBeNull="false" />
     </ElementType>
   </Function>
