+++ Saaloon/Saaloon/Filters/TAccess.cs
@@ -13,10 +13,11 @@ namespace Saaloon.Filters
             //Si Variable de Session is Null, return Login or Type isn't a Teacher
 
             var usuario = HttpContext.Current.Session["Usuario"];
-            string tipo = HttpContext.Current.Session["TipoUsuario"].ToString();
-            if (usuario == null || tipo != "2")
+            var tipo = HttpContext.Current.Session["TipoUsuario"];
+            var tipo2 = HttpContext.Current.Session["Tipo2"];
+            if (usuario == null || tipo != tipo2)
             {
-                filterContext.Result = new RedirectResult("~/Users/LogIn");
+                filterContext.Result = new RedirectResult("~/Home/Login");
             }
             base.OnActionExecuted(filterContext);
         }
