+++ Saaloon/Saaloon/Views/Docente/CrearCurso.cshtml
@@ -1,0 +1,106 @@
+﻿@model Saaloon.Models.CrearCursoDoc
+
+@{
+    ViewBag.Title = "CrearCurso";
+    Layout = "~/Views/Shared/_LayoutDocente.cshtml";
+}
+
+<h2>CrearCurso</h2>
+
+@using (Html.BeginForm("CrearCurso", "Docente", FormMethod.Post, new { @class = "crearalumno" }))
+{
+    @Html.AntiForgeryToken()
+    <br />
+    <center>
+        <div>
+            <img src="~/Content/img/books-pngrepo-com.png" width="130px" />
+        </div>
+        <div class="crearalumno">
+            @Html.ValidationSummary(true, "", new { @class = "text-danger" })
+            <div class="form-row">
+                <div class="col-auto">
+                    <div class="input-group mb-2">
+                        <div class="input-group-prepend">
+                            <div class="input-group-text"><i class="fas fa-book-open"></i></div>
+                        </div>
+                        @Html.EditorFor(model => model.Nombre, new { htmlAttributes = new { @class = "form-control", placeholder = "Nombre" } })
+                        @Html.ValidationMessageFor(model => model.Nombre, "", new { @class = "text-danger" })
+                    </div>
+                </div>
+                <div class="col-auto">
+                    <div class="input-group mb-2">
+                        <div class="input-group-prepend">
+                            <div class="input-group-text"><i class="fas fa-align-center"></i></div>
+                        </div>
+                        @Html.EditorFor(model => model.Descripcion, new { htmlAttributes = new { @class = "form-control", placeholder = "Descripcion" } })
+                        @Html.ValidationMessageFor(model => model.Descripcion, "", new { @class = "text-danger" })
+                    </div>
+                </div>
+            </div>
+            <div class="form-row">
+                <div class="col-auto">
+                    <div class="input-group mb-2">
+                        <div class="input-group-prepend">
+                            <div class="input-group-text"><i class="fas fa-cut"></i></div>
+                        </div>
+                        @Html.EditorFor(model => model.Recursos, new { htmlAttributes = new { @class = "form-control", placeholder = "Recursos" } })
+                        @Html.ValidationMessageFor(model => model.Recursos, "", new { @class = "text-danger" })
+                    </div>
+                </div>
+                <div class="col-auto">
+                    <div class="input-group mb-2">
+                        <div class="input-group-prepend">
+                            <div class="input-group-text"><i class="fas fa-money-bill-wave"></i></div>
+                        </div>
+                        @Html.EditorFor(model => model.Costo, new { htmlAttributes = new { @class = "form-control", placeholder = "Costo" } })
+                        @Html.ValidationMessageFor(model => model.Costo, "", new { @class = "text-danger" })
+                    </div>
+                </div>
+            </div>
+            <div class="form-row">
+                <div class="col-auto">
+                    <div class="input-group mb-2">
+                        <div class="input-group-prepend">
+                            <div class="input-group-text"><i class="fas fa-images"></i></div>
+                        </div>
+                        @Html.EditorFor(model => model.Foto, new { htmlAttributes = new { @class = "form-control", placeholder = "Portada del curso" } })
+                        @Html.ValidationMessageFor(model => model.Foto, "", new { @class = "text-danger" })
+                    </div>
+                </div>
+                <div class="col-auto">
+                    <div class="input-group mb-2">
+                        <div class="input-group-prepend">
+                            <div class="input-group-text"><i class="fas fa-video"></i></div>
+                        </div>
+                        @Html.EditorFor(model => model.VideoIntro, new { htmlAttributes = new { @class = "form-control", placeholder = "Vídeo presentación de curso" } })
+                        @Html.ValidationMessageFor(model => model.VideoIntro, "", new { @class = "text-danger" })
+                    </div>
+                </div>
+            </div>
+            <div class="form-row">
+                <div class="col-auto">
+                    <div class="input-group mb-2">
+                        <div class="input-group-prepend">
+                            <div class="input-group-text"><i class="fas fa-chalkboard-teacher"></i></div>
+                        </div>
+                        @Html.DropDownListFor(x => x.idDocente, (IEnumerable<SelectListItem>)ViewBag.docente, new { @class = "form-control", placeholder = "Docente" })
+                        @Html.ValidationMessageFor(model => model.idDocente, "", new { @class = "text-danger" })
+                    </div>
+                </div>
+            </div>
+
+            @Html.HiddenFor(model => model.idTemario)
+            @Html.HiddenFor(model => model.idDocente)
+            <br />
+            <div class="claseboton">
+                <div>
+                    <div class="btn btn-danger">
+                        @Html.ActionLink("Regresar", "ListaCurso")
+                    </div>
+                    <input type="submit" value="Guardar y continuar" class="btn btn-primary" />
+                </div>
+            </div>
+        </div>
+    </center>
+
+}
