+++ Saaloon/Saaloon/Views/Home/Login.cshtml
@@ -37,8 +37,8 @@
             </div>
         </div>
 
-            <br />
-                <span style="font-size:12px">¿No tienes una cuenta en Virtual Class? </span><a href="Registrarse" style="font-size:12px">Registrate acá </a>
+        <br />
+            <span style="font-size:12px">¿No tienes una cuenta en Virtual Class? </span><a href="Registrarse" style="font-size:12px">Registrate acá </a>
                 <br />
                 <br />
                 <input type="submit" class="btn btn-primary" value="Iniciar Sesión">
