+++ Saaloon/Saaloon/Views/Home/Registrarse.cshtml
@@ -6,7 +6,7 @@
 }
 
 <div class="container">
-    <div  class="card card-container">
+    <div class="card card-container">
         <br />
         <center>
             <h3 class="titleLogin">Registrate</h3>
@@ -25,7 +25,7 @@
                 <div>
                     @Html.TextBoxFor(s => s.Usuario, new { @class = "form-control input-lg", @placeholder = "Usuario" })
                     @Html.ValidationMessageFor(s => s.Usuario, "", new { @class = "text-danger" })
-                </div>    
+                </div>
             </center>
             <br />
             <center>
