+++ Saaloon/Saaloon/Views/Shared/_Layout.cshtml
@@ -25,7 +25,7 @@
                 </li>
             </ul>
             <form class="form-inline my-2 my-lg-0">
-                <a href="@Url.Action("LogIn", "Users")" class="btn btn-outline-primary my-2 my-sm-0">Iniciar Sesion</a>
+                <a href="@Url.Action("Login", "Home")" class="btn btn-outline-primary my-2 my-sm-0">Iniciar Sesion</a>
                 &nbsp;
                 &nbsp;
                 @Html.ActionLink("Registrarse", "Login", "Home", new { @class = "btn btn-outline-success my-2 my-sm-0" })
