+++ Saaloon/Saaloon/Views/Shared/_LayoutDocente.cshtml
@@ -7,6 +7,8 @@
     <script src="https://kit.fontawesome.com/7a98527200.js" crossorigin="anonymous"></script>
     <meta name="viewport" content="width=device-width" />
     <title>@ViewBag.Title</title>
+    @Styles.Render("~/Content/css")
+    @Scripts.Render("~/bundles/modernizr")
 </head>
 <body>
     <div class="navbar navbar-expand-lg navbar-light" style="background-color: #e3f2fd;">
@@ -42,6 +44,7 @@
     <div class="container body-content">
         @RenderBody()
     </div>
+
     <footer class="py-5 bg-dark">
         <div class="container">
             <center>
@@ -55,5 +58,9 @@
 
         </div>
     </footer>
+
+    @Scripts.Render("~/bundles/jquery")
+    @Scripts.Render("~/bundles/bootstrap")
+    @RenderSection("scripts", required: false)
 </body>
 </html>
